Adds front end testing (need to add more tests)
