+++ package.json
@@ -11,28 +11,45 @@
   "scripts": {
     "react-dev": "webpack -d --watch",
     "server-dev": "nodemon server/index.js",
-    "start": "node server/index.js"
+    "start": "node server/index.js",
+    "test": "jest"
   },
   "dependencies": {
     "body-parser": "^1.17.0",
+    "chai": "^3.5.0",
+    "enzyme": "^2.8.2",
     "express": "^4.15.0",
     "highcharts": "^5.0.11",
     "jquery": "^3.2.1",
+    "lodash": "^4.6.1",
     "mysql": "^2.13.0",
     "nodemon": "^1.11.0",
     "react": "^15.4.2",
     "react-bootstrap": "^0.31.0",
-    "react-dom": "^15.4.2"
+    "react-dom": "^15.4.2",
+    "react-test-renderer": "^15.5.4",
+    "require": "^2.4.20",
+    "sinon": "^1.17.3"
   },
   "devDependencies": {
-    "babel-core": "^6.23.1",
-    "babel-loader": "^6.3.2",
     "babel-cli": "^6.7.5",
-    "babel-preset-es2015": "^6.6.0",
+    "babel-core": "^6.23.1",
+    "babel-jest": "^20.0.0",
+    "babel-loader": "^6.3.2",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
     "babel-register": "^6.7.2",
-    "babel-preset-react": "^6.23.0",
-    "webpack": "^2.2.1",
-    "eslint-config-hackreactor": "git://github.com/reactorcore/eslint-config-hackreactor"
+    "chai": "^3.5.0",
+    "enzyme": "^2.8.2",
+    "eslint-config-hackreactor": "git://github.com/reactorcore/eslint-config-hackreactor",
+    "jest": "^20.0.0",
+    "jest-cli": "^20.0.0",
+    "mocha": "2.3.3",
+    "react-addons-test-utils": "^15.5.1",
+    "react-test-renderer": "^15.5.4",
+    "regenerator-runtime": "^0.10.5",
+    "sinon": "1.17.3",
+    "webpack": "^2.2.1"
   },
   "babel": {
     "presets": [
